Improve progress bar accuracy when restarting from a checkpoint (#166)
For steps that have already been trained, we iterate over the sampler as
opposed to the trainloader. This preserves the same sampling
characteristics for a shuffled dataloader while also avoiding the (often
significant) overhead of loading each batch from disk. This PR
re-initializes the progress bar before the actual training continues
from the checkpoint so that the time for iterations over the sampler
aren't averaged with actual training iterations.

Signed-off-by: Matthew Cong <mcong@nvidia.com>