Skip nightly benchmarks if no new commits (#173)
This PR adds a file in the gh-pages benchmark data to store the commits
of fvdb-core and fvdb-reality-capture of the last benchmark run. It also
adds metadata for each run containing these commits.

The workflow is updated to check the current and previous commits and
not run if there are no new commits.

When there are no new commits this finishes within seconds:
https://github.com/openvdb/fvdb-reality-capture/actions/runs/19330065132/job/55290639222

---------

Signed-off-by: Mark Harris <mharris@nvidia.com>